refactored docker command building
