+++ dockercmd/arguments/config_windows.go
@@ -1,0 +1,20 @@
+package arguments
+
+import (
+	"github.com/Oppodelldog/droxy/config"
+	"github.com/Oppodelldog/droxy/dockercmd/builder"
+)
+
+func addGroups(commandDef *config.CommandDefinition, builder *builder.Builder) error {
+	_ = commandDef
+	_ = builder
+
+	return nil
+}
+
+func addImpersonation(commandDef *config.CommandDefinition, builder *builder.Builder) error {
+	_ = commandDef
+	_ = builder
+
+	return nil
+}
