+++ dockercmd/arguments/config_windows_test.go
@@ -1,4 +1,4 @@
-package dockercmd
+package arguments
 
 import (
 	"testing"
