--- dockercmd/config_windows.go
@@ -1,17 +1,0 @@
-package dockercmd
-
-import "github.com/Oppodelldog/droxy/config"
-
-func addGroups(commandDef *config.CommandDefinition, builder *Builder) error {
-	_ = commandDef
-	_ = builder
-
-	return nil
-}
-
-func addImpersonation(commandDef *config.CommandDefinition, builder *Builder) error {
-	_ = commandDef
-	_ = builder
-
-	return nil
-}
